Refactor (APIMF-3491): extracted element finders and pure functions
